dynamically setup invest args at runtime
