+++ backend-worker/worker.py
@@ -30,6 +30,8 @@ from natcap.invest import datastack
 from natcap.invest import utils
 from natcap.invest import validation
 
+import invest_args
+
 logging.basicConfig(level=logging.INFO)
 LOGGER = logging.getLogger(__name__)
 
@@ -90,71 +92,17 @@ NLCD_ORIGIN_X, _, _, NLCD_ORIGIN_Y, _, _ = _NLCD_RASTER_INFO['geotransform']
 PIXELSIZE_X, PIXELSIZE_Y = _NLCD_RASTER_INFO['pixel_size']
 NLCD_COLORS = {}  # update this dict later in file once function is defined.
 
-INVEST_DATA = 'invest-data'
-INVEST_BASE_PATHS = {
-    'docker': f'/opt/appdata/{INVEST_DATA}',
-    'local': os.path.join(
-        os.path.dirname(__file__), '..', 'appdata', INVEST_DATA)
-}
-INVEST_BASE_PATH = None
-for data_path in INVEST_BASE_PATHS.values():
-    if os.path.exists(data_path):
-        INVEST_BASE_PATH = data_path
-        break
-if INVEST_BASE_PATH is None:
-    raise AssertionError(
-        f"Could not find {INVEST_DATA} at any known locations")
-LOGGER.info(f"Using InVEST data at {INVEST_BASE_PATH}")
-
-BIOREGIONS_VECTOR_PATH = f'{INVEST_BASE_PATH}/OE_Bioregions_3857.shp'
-BIOREGIONS_FIELD = 'BIOREGION_'
-# The regions for which we have biophysical parameters
-AVAILABLE_BIOREGIONS = ['NA10', 'NA11', 'NA12', 'NA13', 'NA15', 'NA16', 'NA17',
-                        'NA18', 'NA19', 'NA20', 'NA21', 'NA22', 'NA23', 'NA24',
-                        'NA25', 'NA27', 'NA28', 'NA29', 'NA30', 'NA31', 'NT26']
 CARBON = 'carbon'
 URBAN_COOLING = 'urban_cooling_model'
 INVEST_MODELS = {
-    "pollination": {
-        "api": pollination,
-        "data_key": "pollination",
-        "lulc_args_key": "landcover_raster_path",
-        "args_path": os.path.join(
-            INVEST_BASE_PATH, 'pollination', 'pollination_args.json')},
-    "stormwater": {
-        "api": stormwater,
-        "data_key": "UrbanStormwater",
-        "lulc_args_key": "lulc_path",
-        "args_path": os.path.join(
-            INVEST_BASE_PATH, 'UrbanStormwater', 'urban_stormwater_args.json')},
     URBAN_COOLING: {
         "api": urban_cooling_model,
-        "lulc_args_key": "lulc_raster_path",
-        "aoi_args_key": "aoi_vector_path",
-        "args_path": os.path.join(
-            INVEST_BASE_PATH, URBAN_COOLING,
-            'urban_cooling_model_%s_args.json')
+        "build_args": invest_args.urban_cooling
     },
     CARBON: {
         "api": carbon,
-        "lulc_args_key": "lulc_cur_path",
-        "args_path": os.path.join(
-            INVEST_BASE_PATH, CARBON, 'carbon_%s_args.json'),
-    },
-    "urban_flood_risk_mitigation": {
-        "api": urban_flood_risk_mitigation,
-        "data_key": "UrbanFloodMitigation",
-        "lulc_args_key": "lulc_path",
-        "args_path": os.path.join(
-            INVEST_BASE_PATH, 'UrbanFloodMitigation',
-            'urban_flood_mitigation_args.json')},
-    "urban_nature_access": {
-        "api": urban_nature_access,
-        "data_key": "UrbanNatureAccess",
-        "lulc_args_key": "lulc_raster_path",
-        "args_path": os.path.join(
-            INVEST_BASE_PATH, 'UrbanNatureAccess',
-            'urban_nature_access_args.json')},
+        "build_args": invest_args.carbon
+    }
 }
 LARGEST_SERVICESHED = 2230  # meters https://github.com/natcap/urban-online-workflow/issues/79
 
@@ -782,21 +730,6 @@ def make_thumbnail(pattern_wkt_epsg3857, colors_dict, target_thumbnail_path,
     shutil.rmtree(working_dir, ignore_errors=True)
 
 
-def get_bioregion(point):
-    vector = gdal.OpenEx(BIOREGIONS_VECTOR_PATH, gdal.OF_VECTOR)
-    layer = vector.GetLayer()
-    region = None
-    for feature in layer:
-        geom = shapely.wkt.loads(feature.GetGeometryRef().ExportToWkt())
-        if geom.contains(point):
-            region = feature.GetField(BIOREGIONS_FIELD)
-    if region not in AVAILABLE_BIOREGIONS:
-        raise ValueError(
-            f'the point {point} is not contained '
-            f'within a parameterized bioregion. Region: {region}')
-    return region
-
-
 def do_work(host, port, outputs_location):
     job_queue_url = f'http://{host}:{port}/jobsqueue/'
     LOGGER.info(f'Starting worker, queueing {job_queue_url}')
@@ -920,23 +853,20 @@ def do_work(host, port, outputs_location):
 
                 model_meta = INVEST_MODELS[invest_model]
                 lulc_path = job_args['lulc_source_url']
-                [minx, miny, maxx, maxy] = pygeoprocessing.get_raster_info(
-                    lulc_path)['bounding_box']
-                center = shapely.geometry.Point(
-                    (minx + maxx) / 2, (miny + maxy) / 2)
-                bioregion = get_bioregion(center)
-                args_dict = datastack.extract_parameter_set(
-                    model_meta['args_path'] % bioregion).args
-                args_dict[model_meta['lulc_args_key']] = lulc_path
-                args_dict['workspace_dir'] = os.path.join(
+                
+                workspace_dir = os.path.join(
                     model_outputs_dir, f'{invest_model}-{scenario_id}')
-                LOGGER.info(f'{invest_model} model arguments: {args_dict}')
 
-                # Ultimately we may not need this, but for now this is
-                # convenient for having invest log to a file.
-                with utils.prepare_workspace(args_dict['workspace_dir'],
+                # Ultimately we may not need prepare_workspace, but it is
+                # convenient for 1) creating the workspace as a location to
+                # write dynamically-created input files like an AOI vector, and
+                # 2) having invest log to a file.
+                with utils.prepare_workspace(workspace_dir,
                                              name=invest_model,
                                              logging_level=logging.INFO):
+                    args_dict = model_meta['build_args'](
+                        lulc_path, workspace_dir)
+                    LOGGER.info(f'{invest_model} model arguments: {args_dict}')
                     model_meta['api'].execute(args_dict)
 
                 data = {
